Rename CachedRemoteLayoutStorage → OfflineLayoutStorage
