+++ packages/studio-base/src/services/OfflineLayoutStorage/getNewLayoutName.test.ts
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import getNewLayoutName from "@foxglove/studio-base/services/CachedRemoteLayoutStorage/getNewLayoutName";
+import getNewLayoutName from "./getNewLayoutName";
 
 describe("getNewLayoutName", () => {
   it.each([
