+++ packages/studio-base/src/services/OfflineLayoutStorage/getNewLayoutName.ts
