+++ packages/studio-base/src/services/OfflineLayoutStorage/index.ts
@@ -61,7 +61,7 @@ function getEffectiveMetadata(layout: CachedLayout): LayoutMetadata {
  * This object does not handle any timeout logic and assumes that timeouts from remote storage will
  * be bubbled up as errors.
  */
-export default class CachedRemoteLayoutStorage implements ILayoutStorage {
+export default class OfflineLayoutStorage implements ILayoutStorage {
   /**
    * All access to cache storage is wrapped in a mutex to prevent multi-step operations (such as
    * reading and then writing a single layout, or writing one and deleting another) from getting
